bug fixes, related to needed and optional columns in input files, Version: 0.2.6.9000
Clarification that first column of SV and TS files must be Dummy_ID.
And that Layer_depth_min, Layer_depth_max, Lat_M, and Lon_M are required
by both SV and TS files (not just SV files).  And a workaround in
exploreACMT() for when variables ending in _S and _E are not included.