Update exploreACMT()
CHanged the section that makes the TS plots of # of targets in aTS range and cell to handle a dynamic TS range rather than just a set range as before. It now will make these plots for the range defined in TSrange argument, even when the bins can't be split in 10 dB ranges.