Update estimateLake and Rd file
estimateLake() failed when only one interval density was >0. The failure was actually in a function called by estimateLake() when the function attempts to make a map of interval density with scaled symbols. Symbol scale breaks were not working if there was only one non-zero value. Update now tells the plotting function NOT to plot a variable unless there are at least three intervals with values >0.